organized files a bit more (put config files into a folder)
