+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import App from './App'
-import configureStore from './configureStore'
+import configureStore from './configs/configureStore'
 import * as serviceWorker from './serviceWorker';
 
 const store = configureStore()
